MOB-1725: set automatic default thickness for STL export
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@62813 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/geogebra3D/euclidian3D/openGL/GLBufferManagerTemplatesForPoints.java
@@ -154,7 +154,7 @@ public class GLBufferManagerTemplatesForPoints extends GLBufferManager {
 	 *            point size
 	 */
 	public void selectSphereAndCreateIfNeeded(ManagerShadersElementsGlobalBufferPacking manager,
-			int pointSize) {
+			float pointSize) {
 
 		int templateIndex = ManagerShadersWithTemplates
 				.getIndexForPointSize(pointSize);
